@@ -5,8 +5,8 @@ buildscript {
     ext {
         // Sdk and tools
         minSdkVersion = 16
-        targetSdkVersion = 28
-        compileSdkVersion = 28
+        targetSdkVersion = 29
+        compileSdkVersion = 29
         buildToolsVersion = '27.0.3'
 
         // App dependencies
@@ -21,7 +21,8 @@ buildscript {
         supportLibraryVersion = '1.1.0'
         kotlinVersion = '1.3.31'
         okhttp3='3.4.1'
-
+        jenkinsTalosUrl="$System.env.JENKINS_TALOS_URL"
+        jenkinsLoadboardUrl="$System.env.JENKINS_LOADBOARD_URL"
         deps = [
                 markerClustering: "com.google.maps.android:android-maps-utils:$markerClusterVersion"
         ]
@@ -30,7 +31,6 @@ buildscript {
     repositories {
         google()
         jcenter()
-        mavenCentral()
         maven {
             url 'https://maven.fabric.io/public'
         }
@@ -52,7 +52,15 @@ allprojects {
     repositories {
         google()
         jcenter()
-
+        flatDir {
+            dirs 'libs'
+        }
+    }
+    tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+        kotlinOptions {
+            freeCompilerArgs = ['-Xjvm-default=enable'] //enable or compatibility
+            jvmTarget = "1.8"
+        }
     }
 }
 
