@@ -2,9 +2,11 @@ package com.bykea.pk.partner.ui.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,6 +19,7 @@ import android.widget.Spinner;
 import com.bykea.pk.partner.R;
 import com.bykea.pk.partner.models.data.SignUpAddNumberResponse;
 import com.bykea.pk.partner.models.data.SignUpCity;
+import com.bykea.pk.partner.models.data.SignUpSettingsRecord;
 import com.bykea.pk.partner.models.data.SignUpSettingsResponse;
 import com.bykea.pk.partner.repositories.UserDataHandler;
 import com.bykea.pk.partner.repositories.UserRepository;
@@ -42,6 +45,8 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
+import static com.bykea.pk.partner.utils.Constants.DIGIT_ZERO;
+
 public class NumberRegistration extends Fragment {
     @BindView(R.id.phoneNumberEt)
     FontEditText phoneNumberEt;
@@ -58,6 +63,9 @@ public class NumberRegistration extends Fragment {
 
     @BindView(R.id.tvCity)
     Spinner spCities;
+
+    @BindView(R.id.referenceEt)
+    FontEditText referenceEt;
 
     private RegistrationActivity mCurrentActivity;
     private ArrayList<SignUpCity> mServiceCities = new ArrayList<>();
@@ -111,13 +119,8 @@ public class NumberRegistration extends Fragment {
     }
 
     private void setCitiesAdapter() {
-        SignUpSettingsResponse response = (SignUpSettingsResponse) AppPreferences.getObjectFromSharedPref(SignUpSettingsResponse.class);
-        if (response != null && Utils.isTimeWithInNDay(response.getTimeStamp(), 0.5)) {
-            mCallback.onSignUpSettingsResponse(response);
-        } else {
-            Dialogs.INSTANCE.showLoader(mCurrentActivity);
-            mUserRepository.requestSignUpSettings(mCurrentActivity, mCallback);
-        }
+        Dialogs.INSTANCE.showLoader(mCurrentActivity);
+        mUserRepository.requestSignUpSettings(mCurrentActivity, mCallback);
     }
 
 
@@ -161,16 +164,14 @@ public class NumberRegistration extends Fragment {
         @Override
         public void onSignUpSettingsResponse(final SignUpSettingsResponse response) {
             if (mCurrentActivity != null && getView() != null) {
-                mCurrentActivity.runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        mApiRespinse = response;
-                        mServiceCities.addAll(response.getCity());
-                        dataAdapter1 = new CityDropDownAdapter(mCurrentActivity, mServiceCities);
-                        VIDEO_ID = response.getMain_video();
-                        initAdapter();
-                        Dialogs.INSTANCE.dismissDialog();
-                    }
+                mCurrentActivity.runOnUiThread(() -> {
+                    mApiRespinse = response;
+                    mServiceCities.addAll(response.getCity());
+                    dataAdapter1 = new CityDropDownAdapter(mCurrentActivity, mServiceCities);
+                    VIDEO_ID = response.getMain_video();
+                    initAdapter();
+                    setSignUpDataRecords(response.getRecords());
+                    Dialogs.INSTANCE.dismissDialog();
                 });
             }
         }
@@ -197,6 +198,35 @@ public class NumberRegistration extends Fragment {
             }
         }
     };
+
+    /**
+     * Set SignUp Data Records
+     *
+     * @param records : SignUp Data Records If Received From Server
+     */
+    private void setSignUpDataRecords(SignUpSettingsRecord records) {
+        if (records != null) {
+            if (StringUtils.isNotBlank(records.getCity())) {
+                if (mServiceCities != null && mServiceCities.size() > 0) {
+                    for (int i = DIGIT_ZERO; i < mServiceCities.size(); i++) {
+                        if (StringUtils.isNotEmpty(mServiceCities.get(i).get_id()) &&
+                                mServiceCities.get(i).get_id().equals(records.getCity())) {
+                            if (spCities != null) {
+                                spCities.setSelection(i);
+                            }
+                            break;
+                        }
+                    }
+                }
+            }
+            if (StringUtils.isNotBlank(records.getCnic())) {
+                cnicEt.setText(records.getCnic());
+            }
+            if (StringUtils.isNotBlank(records.getReference())) {
+                referenceEt.setText(records.getReference());
+            }
+        }
+    }
 
     private void logAnalyticsEvent() {
         try {
@@ -240,8 +270,13 @@ public class NumberRegistration extends Fragment {
             case R.id.nextBtn:
                 if (isValidData()) {
                     Dialogs.INSTANCE.showLoader(mCurrentActivity);
-                    mUserRepository.requestRegisterNumber(mCurrentActivity, phoneNumberEt.getText().toString(),
-                            mSelectedCity.get_id(), cnicEt.getText().toString(), mCallback);
+                    mUserRepository
+                            .requestRegisterNumber(mCurrentActivity,
+                                    phoneNumberEt.getText().toString(),
+                                    mSelectedCity.get_id(),
+                                    cnicEt.getText().toString(),
+                                    referenceEt.getText().toString(),
+                                    mCallback);
                 }
                 break;
             case R.id.llIv2:
