@@ -5,10 +5,18 @@ import com.google.gson.annotations.SerializedName;
 
 import org.apache.commons.lang3.StringUtils;
 
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 
 import static com.bykea.pk.partner.utils.Constants.AMOUNT_LIMIT;
 import static com.bykea.pk.partner.utils.Constants.BYKEA_CASH_MAX_AMOUNT;
+import static com.bykea.pk.partner.utils.Constants.DIGIT_ZERO;
+import static com.bykea.pk.partner.utils.Constants.HOW_IT_WORKS_WEB_URL;
+import static com.bykea.pk.partner.utils.Constants.MAX_BATCH_BOOKING_LIMIT;
+import static com.bykea.pk.partner.utils.Constants.MAX_FAHRENHEIT_VALUE;
+import static com.bykea.pk.partner.utils.Constants.MIN_FAHRENHEIT_VALUE;
+import static com.bykea.pk.partner.utils.Constants.OFFLINE_KAMAI_WEB_URL;
 import static com.bykea.pk.partner.utils.Constants.PARTNER_TOP_UP_NEGATIVE_LIMIT_FALLBACK;
 import static com.bykea.pk.partner.utils.Constants.PARTNER_TOP_UP_POSITIVE_LIMIT_FALLBACK;
 
@@ -27,7 +35,17 @@ public class Settings {
 
     @SerializedName("trip_support_max_days_driver")
     private String trip_support_max_days;
-    private String demand;
+
+    /**
+     * this is the configuration flag to manage calls for geocode or nominatum
+     */
+    @SerializedName("is_google_geocode_enable")
+    private boolean isGoogleGeocodeEnable = false;
+    /**
+     * this key will manage whether calculations for easy paisa needs to be run on client or server side
+     */
+    @SerializedName("is_custom_calculations_allow_for_easy_paisa")
+    private Boolean isCustomCalculationsAllowForEasyPaisa;
     private String notice;
     private String top_up_limit;
     private String amount_limit;
@@ -37,8 +55,9 @@ public class Settings {
 
     @SerializedName("terms_driver")
     private String terms;
-    @SerializedName("kronos_partner_summary")
-    private String kronosPartnerSummary;
+
+    @SerializedName("proof_of_delivery_service_codes")
+    private String podServiceCodes;
 
     private String cih_range;
     private String partner_topup_limit;
@@ -66,19 +85,6 @@ public class Settings {
     private boolean withdrawalDisplay;
 
     /**
-     * kronos URL to get booking listings. if null, will starts the older trip flow
-     */
-    @SerializedName("kronos_get_bookings_for_driver")
-    private String bookingLisitingForDriverUrl;
-
-    /**
-     * kronos URL to get booking details by id. if null, will starts the older trip flow
-     */
-    @SerializedName("kronos_get_bookings_by_id")
-    private String bookingDetailByIdUrl;
-
-
-    /**
      * trip fees percentage taken by bykea
      */
     @SerializedName("admin_fee")
@@ -88,7 +94,51 @@ public class Settings {
      * List to get image source url according to priority.
      */
     @SerializedName("priority_list")
-    private HashMap<String,String> priorityList;
+    private HashMap<String, String> priorityList;
+
+    /**
+     * Toggle for ask temperature from partner
+     */
+    @SerializedName("partner_temperature_input_toggle")
+    private boolean partnerTemperatureInputToggle;
+
+    /**
+     * Interval for ask temperature from partner
+     */
+    @SerializedName("partner_temperature_input_interval")
+    private long partnerTemperatureInputInterval;
+
+    /**
+     * Min limit for temperature
+     */
+    @SerializedName("partner_temperature_min_limit")
+    private String partnerTemperatureMinLimit;
+
+    /**
+     * Max limit for temperature
+     */
+    @SerializedName("partner_temperature_max_limit")
+    private String partnerTemperatureMaxLimit;
+    @SerializedName("partner_how_it_works_url")
+    private String howItWorksUrl;
+
+    /**
+     * Max Batch Booking Limit
+     */
+    @SerializedName("batch_booking_limit")
+    private String batchBookingLimit;
+
+    /**
+     * Partner Offline Ride Demo URL
+     */
+    @SerializedName("partner_offline_ride_demo")
+    private String partnerOfflineRideDemo;
+
+    /**
+     * Cool Down Load Board Timer
+     */
+    @SerializedName("cool_down_loadboard_timer")
+    private Long coolDownLoadboardTimer;
 
     public HashMap<String, String> getPriorityList() {
         return priorityList;
@@ -188,11 +238,6 @@ public class Settings {
         return StringUtils.isNotBlank(trip_support_max_days) ? Integer.parseInt(trip_support_max_days) : 10;
     }
 
-    public String getDemand() {
-        return StringUtils.isNotBlank(demand) ? demand : "http://www.bykea.com";
-
-    }
-
     public String getNotice() {
         return StringUtils.isNotBlank(notice) ? notice : "http://www.bykea.com";
 
@@ -280,33 +325,108 @@ public class Settings {
         this.admin_fee = admin_fee;
     }
 
-    public String getBookingLisitingForDriverUrl() {
-        return bookingLisitingForDriverUrl;
+    public List<String> getPodServiceCodes() {
+        return StringUtils.isEmpty(podServiceCodes) ?
+                Arrays.asList(String.valueOf(Constants.ServiceCode.SEND), String.valueOf(Constants.ServiceCode.SEND_COD),
+                        String.valueOf(Constants.ServiceCode.OFFLINE_DELIVERY)) : Arrays.asList(podServiceCodes.split("\\s*,\\s*"));
     }
 
-    public void setBookingLisitingForDriverUrl(String bookingLisitingForDriverUrl) {
-        this.bookingLisitingForDriverUrl = bookingLisitingForDriverUrl;
-    }
-
-    public String getBookingDetailByIdUrl() {
-        return bookingDetailByIdUrl;
-    }
-
-    public void setBookingDetailByIdUrl(String bookingDetailByIdUrl) {
-        this.bookingDetailByIdUrl = bookingDetailByIdUrl;
+    public void setPodServiceCodes(String podServiceCodes) {
+        this.podServiceCodes = podServiceCodes;
     }
 
     /**
-     * @return will return the summary api url of kronos
+     * Check whether partner temperature feature is enable or not
+     *
+     * @return true, if ask temperature else false
      */
-    public String getKronosPartnerSummary() {
-        return kronosPartnerSummary;
+    public boolean isPartnerTemperatureInputToggle() {
+        return partnerTemperatureInputToggle;
     }
 
     /**
-     * @return will set the summary api url of kronos
+     * Getter for time interval (seconds) to ask input for temperature after this interval
+     *
+     * @return long: Time interval in seconds
      */
-    public void setKronosPartnerSummary(String kronosPartnerSummary) {
-        this.kronosPartnerSummary = kronosPartnerSummary;
+    public long getPartnerTemperatureInputInterval() {
+        return partnerTemperatureInputInterval;
+    }
+
+    public double getPartnerTemperatureMinLimit() {
+        try {
+            return Double.parseDouble(partnerTemperatureMinLimit);
+        } catch (Exception e) {
+            return MIN_FAHRENHEIT_VALUE;
+        }
+    }
+
+    public double getPartnerTemperatureMaxLimit() {
+        try {
+            return Double.parseDouble(partnerTemperatureMaxLimit);
+        } catch (Exception e) {
+            return MAX_FAHRENHEIT_VALUE;
+        }
+    }
+
+    public int getBatchBookingLimit() {
+        try {
+            return Integer.parseInt(batchBookingLimit);
+        } catch (Exception e) {
+            return MAX_BATCH_BOOKING_LIMIT;
+        }
+    }
+
+    /**
+     * this key will manage whether calculations for easy paisa needs to be run on client or server side
+     */
+    public boolean isCustomCalculationsAllowForEasyPaisa() {
+        return isCustomCalculationsAllowForEasyPaisa == null ? true : isCustomCalculationsAllowForEasyPaisa;
+    }
+
+    /**
+     * this key will manage whether calculations for easy paisa needs to be run on client or server side
+     */
+    public void setCustomCalculationsAllowForEasyPaisa(boolean customCalculationsAllowForEasyPaisa) {
+        isCustomCalculationsAllowForEasyPaisa = customCalculationsAllowForEasyPaisa;
+    }
+
+    public String getHowItWorksUrl() {
+        return StringUtils.isEmpty(howItWorksUrl) ? HOW_IT_WORKS_WEB_URL : howItWorksUrl;
+    }
+
+    public void setHowItWorksUrl(String howItWorksUrl) {
+        this.howItWorksUrl = howItWorksUrl;
+    }
+
+    /**
+     * Get Offline Kamari Web Url
+     *
+     * @return String: If not recevied in setting return fallback
+     */
+    public String getPartnerOfflineRideDemo() {
+        return StringUtils.isEmpty(partnerOfflineRideDemo) ? OFFLINE_KAMAI_WEB_URL : partnerOfflineRideDemo;
+    }
+
+    /**
+     * this is the configuration flag to manage calls for geocode or nominatum
+     */
+    public boolean isGoogleGeocodeEnable() {
+        return isGoogleGeocodeEnable;
+    }
+
+    /**
+     * this is the configuration flag to manage calls for geocode or nominatum
+     */
+    public void setGoogleGeocodeEnable(boolean googleGeocodeEnable) {
+        isGoogleGeocodeEnable = googleGeocodeEnable;
+    }
+    /**
+     * Cool Down Load Board Timer
+     *
+     * @return if exit else null
+     */
+    public Long getCoolDownLoadboardTimer() {
+        return coolDownLoadboardTimer == null ? DIGIT_ZERO : coolDownLoadboardTimer;
     }
 }
