@@ -1,8 +1,8 @@
 package com.bykea.pk.partner.repositories.places;
 
 import com.bykea.pk.partner.models.data.PlacesList;
+import com.bykea.pk.partner.models.response.BykeaDistanceMatrixResponse;
 import com.bykea.pk.partner.models.response.GeoCodeApiResponse;
-import com.bykea.pk.partner.models.response.GoogleDistanceMatrixApi;
 import com.bykea.pk.partner.models.response.PlaceAutoCompleteResponse;
 import com.bykea.pk.partner.models.response.PlaceDetailsResponse;
 
@@ -17,7 +17,7 @@ public class PlacesDataHandler implements IPlacesDataHandler {
     }
 
     @Override
-    public void onDistanceMatrixResponse(GoogleDistanceMatrixApi response) {
+    public void onDistanceMatrixResponse(BykeaDistanceMatrixResponse response) {
 
     }
 
