@@ -5,8 +5,8 @@ import android.content.Context;
 import com.bykea.pk.partner.communication.IResponseCallback;
 import com.bykea.pk.partner.communication.rest.RestRequestHandler;
 import com.bykea.pk.partner.communication.socket.WebIORequestHandler;
+import com.bykea.pk.partner.models.response.BykeaDistanceMatrixResponse;
 import com.bykea.pk.partner.models.response.GeoCodeApiResponse;
-import com.bykea.pk.partner.models.response.GoogleDistanceMatrixApi;
 import com.bykea.pk.partner.models.response.PlaceAutoCompleteResponse;
 import com.bykea.pk.partner.models.response.PlaceDetailsResponse;
 import com.bykea.pk.partner.ui.helpers.AppPreferences;
@@ -33,11 +33,11 @@ public class PlacesRepository {
         public void onResponse(Object object) {
             if (object instanceof String) {
                 mUserCallback.onPlacesResponse((String) object);
-            } else if (object instanceof GoogleDistanceMatrixApi) {
-                mUserCallback.onDistanceMatrixResponse((GoogleDistanceMatrixApi) object);
+            } else if (object instanceof BykeaDistanceMatrixResponse) {
+                mUserCallback.onDistanceMatrixResponse((BykeaDistanceMatrixResponse) object);
             } else if (object instanceof PlaceDetailsResponse) {
                 mUserCallback.onPlaceDetailsResponse((PlaceDetailsResponse) object);
-            }else if (object instanceof PlaceAutoCompleteResponse) {
+            } else if (object instanceof PlaceAutoCompleteResponse) {
                 mUserCallback.onPlaceAutoCompleteResponse((PlaceAutoCompleteResponse) object);
             } else if (object instanceof GeoCodeApiResponse) {
                 mUserCallback.onGeoCodeApiResponse((GeoCodeApiResponse) object);
@@ -52,16 +52,15 @@ public class PlacesRepository {
     };
 
     /**
-     *
-     * @param origin : Longitude and Latitude - String Concatenation
+     * @param origin      : Longitude and Latitude - String Concatenation
      * @param destination : Longitude and Latitude - String Concatenation
-     * @param context : Calling Activity
-     * @param handler : CallBack
+     * @param context     : Calling Activity
+     * @param handler     : CallBack
      */
 
     public void getDistanceMatrix(String origin, String destination, Context context, IPlacesDataHandler handler) {
         mUserCallback = handler;
-        mRestRequestHandler.getDistanceMatriax(origin, destination, mGeoCoderPlaces, context);
+        mRestRequestHandler.getDistanceMatrix(origin, destination, mGeoCoderPlaces, context);
     }
 
     /**
@@ -76,7 +75,7 @@ public class PlacesRepository {
 
     /**
      * @param context : Calling Activity
-     * @param search : Filter String Value
+     * @param search  : Filter String Value
      * @param handler : CallBack
      */
     public void getPlaceAutoComplete(Context context, String search, IPlacesDataHandler handler) {
@@ -93,5 +92,11 @@ public class PlacesRepository {
         mUserCallback = handler;
         AppPreferences.setGeoCoderApiKeyRequired(true);
         mRestRequestHandler.callGeoCodeApiWithPlaceId(placeId, context, mGeoCoderPlaces);
+    }
+
+    public void getOSMGeoCoder(IPlacesDataHandler handler,
+                               String lat, String lng, Context context) {
+        mUserCallback = handler;
+        mRestRequestHandler.callOSMGeoCoderApi(lat, lng, mGeoCoderPlaces, context);
     }
 }
