@@ -1,25 +1,36 @@
 package com.bykea.pk.partner.ui.common
 
+import android.graphics.Color
+import android.graphics.Paint
+import android.graphics.Paint.STRIKE_THRU_TEXT_FLAG
 import android.view.View
 import android.widget.ImageView
 import android.widget.ListView
+import android.widget.TextView
+import androidx.appcompat.widget.AppCompatImageView
 import androidx.core.content.ContextCompat
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.bykea.pk.partner.DriverApp
 import com.bykea.pk.partner.R
 import com.bykea.pk.partner.dal.Job
 import com.bykea.pk.partner.dal.Rules
+import com.bykea.pk.partner.dal.source.remote.request.nodataentry.DeliveryDetails
 import com.bykea.pk.partner.dal.util.SEPERATOR
+import com.bykea.pk.partner.ui.helpers.FontUtils
 import com.bykea.pk.partner.ui.loadboard.list.JobListAdapter
 import com.bykea.pk.partner.utils.Constants
 import com.bykea.pk.partner.utils.Constants.*
 import com.bykea.pk.partner.utils.Constants.ServiceCode.*
+import com.bykea.pk.partner.utils.TripStatus.*
+import com.bykea.pk.partner.utils.Util
 import com.bykea.pk.partner.utils.Utils
 import com.bykea.pk.partner.widgets.AutoFitFontTextView
 import com.bykea.pk.partner.widgets.FontTextView
 import org.apache.commons.lang3.StringUtils
 import java.text.SimpleDateFormat
 import java.util.*
+import kotlin.collections.ArrayList
 
 
 /**
@@ -31,10 +42,44 @@ object BindingAdapters {
 
     @BindingAdapter("app:items")
     @JvmStatic
-    fun setItems(recyclerView: RecyclerView, list: List<Any>) {
-        with(recyclerView.adapter as LastAdapter<Any>) {
-            items = list
+    fun setItems(recyclerView: RecyclerView, list: List<Any>?) {
+        list?.let {
+            with(recyclerView.adapter as LastAdapter<Any>) {
+                items = ArrayList(it)
+            }
         }
+    }
+
+    @BindingAdapter("app:hexTextColor")
+    @JvmStatic
+    fun setHexTextColor(textView: TextView, color: String?) {
+        textView.apply {
+            setTextColor(if (StringUtils.isNotEmpty(color)) {
+                Color.parseColor(color)
+            } else {
+                Color.BLACK
+            })
+        }
+    }
+
+    @BindingAdapter("app:fontType")
+    @JvmStatic
+    fun setFontType(textView: TextView, type: String?) {
+        type?.let {
+            if (type.equals("ur", false))
+                textView.typeface =
+                        FontUtils.getFonts(textView.context, "jameel_noori_nastaleeq.ttf")
+        }
+    }
+
+    @BindingAdapter("app:activeRadio")
+    @JvmStatic
+    fun activeRadio(imageView: ImageView, isActive: Boolean) {
+        imageView.setImageResource(if (isActive) {
+            R.drawable.selected_radio_button
+        } else {
+            R.drawable.unselected_radio_button
+        })
     }
 
     @BindingAdapter("app:items")
@@ -42,6 +87,16 @@ object BindingAdapters {
     fun setItems(listView: ListView, items: List<Job>) {
         with(listView.adapter as JobListAdapter) {
             replaceData(items)
+        }
+    }
+
+    @BindingAdapter("app:showLineOverText")
+    @JvmStatic
+    fun showStrikeLine(textView: TextView, enable: Boolean) {
+        textView.paintFlags = if (enable) {
+            textView.paintFlags or STRIKE_THRU_TEXT_FLAG
+        } else {
+            textView.paintFlags and Paint.STRIKE_THRU_TEXT_FLAG.inv()
         }
     }
 
@@ -55,17 +110,24 @@ object BindingAdapters {
         Utils.loadImgPicasso(imageView, R.color.white, url)
     }
 
+    @BindingAdapter("app:urduWrappingText")
+    @JvmStatic
+    fun urduWrappingText(textView: TextView, url: String? = null) {
+        if (url == null) return
+        textView.text = String.format(textView.resources.getString(R.string.empty_string_sandwich_placeholder), url)
+    }
+
     @BindingAdapter("app:serviceCode")
     @JvmStatic
     fun setServiceCode(imageView: ImageView, serviceCode: Int) {
         when (serviceCode) {
-            SEND -> imageView.setImageResource(R.drawable.bhejdo_no_caption)
-            SEND_COD -> imageView.setImageResource(R.drawable.bhejdo_no_caption)
+            SEND, SEND_COD, MULTI_DELIVERY, NEW_BATCH_DELIVERY -> imageView.setImageResource(R.drawable.bhejdo_no_caption)
             RIDE, DISPATCH_RIDE -> imageView.setImageResource(R.drawable.ride_right)
             MART -> imageView.setImageResource(R.drawable.ic_purchase)
             COURIER -> imageView.setImageResource(R.drawable.courier_no_caption)
             MOBILE_TOP_UP -> imageView.setImageResource(R.drawable.ic_pay)
             MOBILE_WALLET -> imageView.setImageResource(R.drawable.ic_pay)
+            FOOD -> imageView.setImageResource(R.drawable.ic_food)
             BANK_TRANSFER -> imageView.setImageResource(R.drawable.ic_pay)
             UTILITY -> imageView.setImageResource(R.drawable.ic_pay)
             else -> imageView.setImageResource(R.drawable.ride_right)
@@ -76,6 +138,23 @@ object BindingAdapters {
     @JvmStatic
     fun setGoneUnless(view: View, visible: Boolean) {
         view.visibility = if (visible) View.VISIBLE else View.GONE
+    }
+
+    @BindingAdapter("app:invisibleUnless")
+    @JvmStatic
+    fun setInvisibleUnless(view: View, value: Any? = null) {
+        value?.let {
+            when (value) {
+                is Boolean -> {
+                    view.visibility = if (value) View.VISIBLE else View.INVISIBLE
+                }
+                is String -> {
+                    view.visibility = if (value.isNotEmpty()) View.VISIBLE else View.INVISIBLE
+                }
+            }
+        } ?: run {
+            view.visibility = View.INVISIBLE
+        }
     }
 
     @BindingAdapter("app:dateFormat")
@@ -169,8 +248,65 @@ object BindingAdapters {
                 }
             }
         } else {
-            autoFontTextView.setTextColor(ContextCompat.getColor(autoFontTextView.context,  R.color.textColorPerformance))
+            autoFontTextView.setTextColor(ContextCompat.getColor(autoFontTextView.context, R.color.textColorPerformance))
             autoFontTextView.setBackgroundResource(R.drawable.gray_left_top_bottom_bordered_bg)
         }
     }
+
+    @BindingAdapter("app:amountFormatted")
+    @JvmStatic
+    fun setAmountFormatted(fontTextView: FontTextView, amount: Any? = null) {
+        amount?.let {
+            when (it) {
+                is Int -> {
+                    fontTextView.text = String.format(DriverApp.getContext().getString(R.string.amount_rs), it.toString())
+                }
+                is String -> {
+                    fontTextView.text = String.format(DriverApp.getContext().getString(R.string.amount_rs), it)
+                }
+            }
+        }
+    }
+
+    @BindingAdapter("app:batchTripItemAccordingToStatus")
+    @JvmStatic
+    fun setBatchTripItemAccordingToStatus(view: View, deliveryDetails: DeliveryDetails?) {
+        val imageViewShowDetails: AppCompatImageView = view.findViewById(R.id.imageViewShowDetails)
+        val imageViewEdit: AppCompatImageView = view.findViewById(R.id.imageViewEdit)
+        val textViewStatus: FontTextView = view.findViewById(R.id.textViewStatus)
+
+        Util.safeLet(deliveryDetails,
+                deliveryDetails?.details,
+                deliveryDetails?.details?.status) { _, _, status ->
+            when (status.toLowerCase()) {
+                ON_START_TRIP.toLowerCase() -> {
+                    imageViewShowDetails.visibility = View.VISIBLE
+                    imageViewEdit.visibility = View.INVISIBLE
+                    textViewStatus.visibility = View.GONE
+                }
+                ON_FEEDBACK_TRIP.toLowerCase(), ON_COMPLETED_TRIP.toLowerCase(), ON_FINISH_TRIP.toLowerCase() -> {
+                    imageViewShowDetails.visibility = View.INVISIBLE
+                    imageViewEdit.visibility = View.INVISIBLE
+                    deliveryDetails?.details?.delivery_status?.let {
+                        textViewStatus.visibility = View.VISIBLE
+                        if (it) {
+                            textViewStatus.text = DriverApp.getContext().getText(R.string.successful)
+                            textViewStatus.setTextColor(ContextCompat.getColor(DriverApp.getContext(), R.color.color_blue_fleet))
+                        } else {
+                            textViewStatus.text = DriverApp.getContext().getText(R.string.unsuccessful)
+                            textViewStatus.setTextColor(ContextCompat.getColor(DriverApp.getContext(), R.color.booking_red))
+                        }
+                    } ?: run {
+                        textViewStatus.visibility = View.GONE
+                    }
+                }
+                else -> {
+                    imageViewShowDetails.visibility = View.VISIBLE
+                    imageViewEdit.visibility = View.VISIBLE
+                    textViewStatus.visibility = View.GONE
+                }
+            }
+        }
+    }
+
 }
