@@ -7,6 +7,7 @@ import android.util.Log;
 import com.bykea.pk.partner.BuildConfig;
 import com.bykea.pk.partner.DriverApp;
 import com.bykea.pk.partner.dal.LocCoordinatesInTrip;
+import com.bykea.pk.partner.dal.source.remote.response.DriverSettingsResponse;
 import com.bykea.pk.partner.models.ReceivedMessageCount;
 import com.bykea.pk.partner.models.data.CitiesData;
 import com.bykea.pk.partner.models.data.MultiDeliveryCallDriverData;
@@ -19,7 +20,7 @@ import com.bykea.pk.partner.models.data.TrackingData;
 import com.bykea.pk.partner.models.data.ZoneData;
 import com.bykea.pk.partner.models.response.GetCitiesResponse;
 import com.bykea.pk.partner.models.response.NormalCallData;
-import com.bykea.pk.partner.models.data.MultiDeliveryCallDriverData;
+import com.bykea.pk.partner.models.response.RegistrationLinksToken;
 import com.bykea.pk.partner.models.response.ZoneAreaResponse;
 import com.bykea.pk.partner.utils.Constants;
 import com.bykea.pk.partner.utils.Keys;
@@ -34,6 +35,9 @@ import org.apache.commons.lang3.StringUtils;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Iterator;
+import java.util.Map;
+
+import static com.bykea.pk.partner.utils.Constants.DIGIT_ZERO;
 
 public class AppPreferences {
 
@@ -46,11 +50,45 @@ public class AppPreferences {
                 .apply();
     }
 
+    /**
+     * Clear Particulars Shared Preferences
+     * (Which are not required to preserve)
+     */
+    public static void clearExceptParticulars() {
+        for (String key : mSharedPreferences.getAll().keySet()) {
+            if (!Utils.isSharedPreferenceKeyPreserveRequired(key)) {
+                mSharedPreferences.edit().remove(key).apply();
+            }
+        }
+    }
+
     public static void saveSettingsData(SettingsData data) {
         mSharedPreferences
                 .edit()
                 .putString(Keys.SETTING_DATA, new Gson().toJson(data))
                 .apply();
+    }
+
+    /**
+     * Save Driver Settings Data
+     * @param data
+     */
+    public static void saveDriverSettingsData(DriverSettingsResponse data) {
+        mSharedPreferences
+                .edit()
+                .putString(Keys.DRIVER_SETTINGS, new Gson().toJson(data))
+                .apply();
+    }
+
+    public static DriverSettingsResponse getDriverSettings() {
+        String data = mSharedPreferences.getString(Keys.DRIVER_SETTINGS, StringUtils.EMPTY);
+        DriverSettingsResponse settingsData;
+        if (StringUtils.isBlank(data)) {
+            settingsData = new DriverSettingsResponse();
+        } else {
+            settingsData = new Gson().fromJson(data, DriverSettingsResponse.class);
+        }
+        return settingsData;
     }
 
 
@@ -550,6 +588,17 @@ public class AppPreferences {
         return mSharedPreferences.getString(Keys.ADMIN_MSG, StringUtils.EMPTY);
     }
 
+    public static void setListDeliveryActivityOnForeground(boolean value) {
+        mSharedPreferences
+                .edit()
+                .putBoolean(Keys.LIST_DELIVERY_ACTIVITY_FOREGROUND, value)
+                .apply();
+    }
+
+    public static boolean isListDeliveryOnForeground() {
+        return mSharedPreferences.getBoolean(Keys.LIST_DELIVERY_ACTIVITY_FOREGROUND, false);
+    }
+
     public static boolean isJobActivityOnForeground() {
         return mSharedPreferences.getBoolean(Keys.JOB_ACTIVITY_FOREGROUND, false);
     }
@@ -571,6 +620,7 @@ public class AppPreferences {
 
     /**
      * Set Mutlidelivery activity is visible or not
+     *
      * @param value : True or False
      */
     public static void setMultiDeliveryJobActivityOnForeground(boolean value) {
@@ -583,6 +633,7 @@ public class AppPreferences {
 
     /**
      * Check If Distance Matrix Is Required To Call For Mutlidelivery Or Not
+     *
      * @return true if required else false
      */
     public static boolean isMultiDeliveryDistanceMatrixCalledRequired() {
@@ -591,6 +642,7 @@ public class AppPreferences {
 
     /**
      * Set Distance Matrix Is Required To Call For Mutlidelivery Or Not
+     *
      * @param value : True or False
      */
     public static void setMultiDeliveryDistanceMatrixCalledRequired(boolean value) {
@@ -602,7 +654,8 @@ public class AppPreferences {
 
     /**
      * Check If Drop Of Update Is Required Or Not
-     * @return  True/False
+     *
+     * @return True/False
      */
     public static boolean isDropOffUpdateRequired() {
         return mSharedPreferences.getBoolean(Keys.DROP_OFF_UPDATE_REQUIRED, false);
@@ -612,6 +665,7 @@ public class AppPreferences {
      * Set Drop Off Update Required
      * If Application Is Running But Booking Activity Is In Background or Chat Activity Is In Foreground
      * Set To False After Execution Of True
+     *
      * @param value : True/False (On Behalf Of Above Decision)
      */
     public static void setDropOffUpdateRequired(boolean value) {
@@ -852,7 +906,6 @@ public class AppPreferences {
         prevLatLngList.add(currentLatLng);
         String value = new Gson().toJson(prevLatLngList, new TypeToken<ArrayList<LocCoordinatesInTrip>>() {
         }.getType());
-        Utils.redLog("InTripLoc", value);
         mSharedPreferences.edit().putString(Keys.IN_TRIP_LAT_LNG_ARRAY, value).apply();
     }
 
@@ -868,7 +921,6 @@ public class AppPreferences {
         prevLatLngList.add(currentLatLng);
         String value = new Gson().toJson(prevLatLngList, new TypeToken<ArrayList<LocCoordinatesInTrip>>() {
         }.getType());
-        Utils.redLog("InTripLoc", value);
         mSharedPreferences.edit().putString(Keys.IN_TRIP_LAT_LNG_ARRAY, value).apply();
     }
 
@@ -878,7 +930,6 @@ public class AppPreferences {
             prevLatLngList.addAll(index, routeLatLngList);
             String value = new Gson().toJson(prevLatLngList, new TypeToken<ArrayList<LocCoordinatesInTrip>>() {
             }.getType());
-            Utils.redLog("InTripLoc", value);
             mSharedPreferences.edit().putString(Keys.IN_TRIP_LAT_LNG_ARRAY, value).apply();
         }
     }
@@ -1623,6 +1674,7 @@ public class AppPreferences {
 
     /**
      * Set Booking Voice Note Url Available
+     *
      * @param voiceNoteUrl : Voice Note Url
      */
     public static void setBookingVoiceNoteUrlAvailable(String voiceNoteUrl) {
@@ -1638,5 +1690,75 @@ public class AppPreferences {
         SharedPreferences.Editor ed = mSharedPreferences.edit();
         ed.remove(Keys.BOOKING_VOICE_NOTE_URL);
         ed.commit();
+    }
+
+    /**
+     * Get last partner submit time.
+     *
+     * @return long : Time in milliseconds
+     */
+    public static long getLastPartnerTemperatureSubmitTime() {
+        return mSharedPreferences.getLong(Keys.LAST_PARTNER_TEMPERATURE_SUBMIT, DIGIT_ZERO);
+    }
+
+    /**
+     * Set last partner submit time.
+     *
+     * @param value : Time in milliseconds
+     */
+    public static void setLastPartnerTemperatureSubmitTime(long value) {
+        mSharedPreferences
+                .edit()
+                .putLong(Keys.LAST_PARTNER_TEMPERATURE_SUBMIT, value)
+                .apply();
+    }
+
+    public static void setLastSelectedMsgPosition(int position, String message) {
+        mSharedPreferences
+                .edit()
+                .putInt(Keys.LAST_SELECTED_MSG_POSITION, position)
+                .apply();
+        if (message != null) {
+            mSharedPreferences
+                    .edit()
+                    .putString(Keys.LAST_SELECTED_FEEDBACK_MSG, message)
+                    .apply();
+        }
+    }
+
+    public static int getLastSelectedMsgPosition() {
+        return mSharedPreferences.getInt(Keys.LAST_SELECTED_MSG_POSITION, DIGIT_ZERO);
+    }
+
+    public static String getLastSelectedMsg() {
+        return mSharedPreferences.getString(Keys.LAST_SELECTED_FEEDBACK_MSG, StringUtils.EMPTY);
+    }
+
+    /**
+     * Save RegistrationLinksToken in shared preference.
+     *
+     * @param response The {@link RegistrationLinksToken} object.
+     */
+    public static void setRegistrationLinksToken(RegistrationLinksToken response) {
+        mSharedPreferences
+                .edit()
+                .putString(Keys.SIGNUP_LINKS_TOKENS, new Gson().toJson(response))
+                .apply();
+    }
+
+    /**
+     * Fetch the RegistrationLinksToken from shared preference.
+     *
+     * @return The {@link RegistrationLinksToken} object.
+     */
+    public static RegistrationLinksToken getRegistrationLinksToken() {
+        Gson gson = new Gson();
+        try {
+            return gson.fromJson(
+                    mSharedPreferences.getString(Keys.SIGNUP_LINKS_TOKENS, StringUtils.EMPTY),
+                    RegistrationLinksToken.class);
+        } catch (Exception e) {
+            return null;
+        }
     }
 }
