@@ -12,8 +12,13 @@ import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
 import com.bykea.pk.partner.R
 import com.bykea.pk.partner.analytics.AnalyticsEventsJsonObjects
+import com.bykea.pk.partner.dal.Trips
 import com.bykea.pk.partner.databinding.JobDetailActBinding
+import com.bykea.pk.partner.models.data.MultiDeliveryCallDriverData
+import com.bykea.pk.partner.models.response.CheckDriverStatusResponse
+import com.bykea.pk.partner.repositories.UserDataHandler
 import com.bykea.pk.partner.ui.activities.BaseActivity
+import com.bykea.pk.partner.ui.common.LastAdapter
 import com.bykea.pk.partner.ui.common.obtainViewModel
 import com.bykea.pk.partner.ui.common.setupSnackbar
 import com.bykea.pk.partner.ui.helpers.ActivityStackManager
@@ -31,6 +36,8 @@ import com.google.android.gms.maps.MapsInitializer
 import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.*
 import com.google.android.material.snackbar.Snackbar
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import kotlinx.android.synthetic.main.activity_confirm_drop_off_address.*
 import kotlinx.android.synthetic.main.job_detail_act.*
 import java.io.File
@@ -51,8 +58,9 @@ class JobDetailActivity : BaseActivity() {
     private lateinit var mGoogleMap: GoogleMap
     private var driverMarker: Marker? = null
 
-    private var mediaPlayer: MediaPlayer? = null;
-    private val handler: Handler = Handler();
+    private var mediaPlayer: MediaPlayer? = null
+    private val handler: Handler = Handler()
+    private var lastAdapter: LastAdapter<Trips>? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -76,7 +84,30 @@ class JobDetailActivity : BaseActivity() {
                 } else {
                     AppPreferences.removeBookingVoiceNoteUrl()
                 }
-                ActivityStackManager.getInstance().startJobActivity(this@JobDetailActivity, false)
+                if (Utils.isNewBatchService(binding.viewmodel?.job?.value?.service_code!!)) {
+                    ActivityStackManager.getInstance().startJobActivity(this@JobDetailActivity, false)
+                } else if (binding.viewmodel?.job?.value?.trip_type?.equals(Constants.TripTypes.BATCH_TYPE, true)!!) {
+                    Dialogs.INSTANCE.showLoader(this@JobDetailActivity)
+                    Handler().postDelayed({
+                        binding.viewmodel?.getTripDetails(object : UserDataHandler() {
+                            override fun onRunningTrips(response: CheckDriverStatusResponse?) {
+                                super.onRunningTrips(response)
+                                Dialogs.INSTANCE.dismissDialog()
+                                if (response?.isSuccess()!!) {
+                                    val trip = Gson().toJson(response.data.trip)
+                                    val type = object : TypeToken<MultiDeliveryCallDriverData>() {
+                                    }.type
+                                    AppPreferences.setDeliveryType(Constants.CallType.BATCH)
+                                    val deliveryCallDriverData: MultiDeliveryCallDriverData = Gson().fromJson(trip, type)
+                                    AppPreferences.setMultiDeliveryCallDriverData(deliveryCallDriverData)
+                                    ActivityStackManager.getInstance().startMultiDeliveryBookingActivity(this@JobDetailActivity)
+                                }
+                            }
+                        })
+                    }, 3000)
+                } else {
+                    ActivityStackManager.getInstance().startJobActivity(this@JobDetailActivity, false)
+                }
             })
 
             bookingTakenCommand.observe(this@JobDetailActivity, Observer {
@@ -85,6 +116,14 @@ class JobDetailActivity : BaseActivity() {
                                 this@JobDetailActivity.resources.getString(R.string.booking_already_taken_title),
                                 this@JobDetailActivity.resources.getString(R.string.booking_already_taken_msg)
                         ) { finish() }
+            })
+            driverBlockedByAdmin.observe(this@JobDetailActivity, Observer {
+                if (it) {
+                    binding.viewmodel?.driverBlockedByAdmin?.value = false
+                    Dialogs.INSTANCE.showRegionOutErrorDialog(this@JobDetailActivity,
+                            Utils.getSupportHelplineNumber(),
+                            getString(R.string.account_blocked_wallet_amount_not_paid))
+                }
             })
         }
         binding.listener = object : JobDetailActionsListener {
@@ -152,6 +191,9 @@ class JobDetailActivity : BaseActivity() {
             }
         }
         AppPreferences.removeReceivedMessageCount();
+
+        lastAdapter = LastAdapter(R.layout.list_item_trips)
+        recViewTrips.adapter = lastAdapter
     }
 
     /**
@@ -169,27 +211,32 @@ class JobDetailActivity : BaseActivity() {
             startPlayProgressUpdater()
         } else {
             Dialogs.INSTANCE.showLoader(this@JobDetailActivity)
-            BykeaAmazonClient.getFileObject(url, object : Callback<File> {
-                override fun success(obj: File) {
-                    Dialogs.INSTANCE.dismissDialog()
+            AppPreferences.getDriverSettings()?.data?.s3BucketVoiceNotes?.let {
+                BykeaAmazonClient.getFileObject(url, object : Callback<File> {
+                    override fun success(obj: File) {
+                        Dialogs.INSTANCE.dismissDialog()
 
-                    imgViewAudioPlay.setImageDrawable(resources.getDrawable(R.drawable.ic_audio_stop))
-                    imgViewAudioPlay.isEnabled = false
-                    progressBarForAudioPlay.visibility = View.VISIBLE
+                        imgViewAudioPlay.setImageDrawable(resources.getDrawable(R.drawable.ic_audio_stop))
+                        imgViewAudioPlay.isEnabled = false
+                        progressBarForAudioPlay.visibility = View.VISIBLE
 
-                    mediaPlayer = MediaPlayer()
-                    mediaPlayer?.setDataSource(FileInputStream(obj).fd);
-                    mediaPlayer?.prepare()
-                    progressBarForAudioPlay.setMax(mediaPlayer?.duration!!);
-                    mediaPlayer?.start()
-                    startPlayProgressUpdater()
-                }
+                        mediaPlayer = MediaPlayer()
+                        mediaPlayer?.setDataSource(FileInputStream(obj).fd);
+                        mediaPlayer?.prepare()
+                        progressBarForAudioPlay.setMax(mediaPlayer?.duration!!);
+                        mediaPlayer?.start()
+                        startPlayProgressUpdater()
+                    }
 
-                override fun fail(errorCode: Int, errorMsg: String) {
-                    Dialogs.INSTANCE.dismissDialog()
-                    binding.viewmodel?.showSnackbarMessage(R.string.no_voice_note_available)
-                }
-            })
+                    override fun fail(errorCode: Int, errorMsg: String) {
+                        Dialogs.INSTANCE.dismissDialog()
+                        binding.viewmodel?.showSnackbarMessage(R.string.no_voice_note_available)
+                    }
+                }, it)
+            } ?: run {
+                Dialogs.INSTANCE.dismissDialog()
+                Dialogs.INSTANCE.showToast(getString(R.string.settings_are_not_updated))
+            }
         }
     }
 
