@@ -11,13 +11,19 @@ import com.bykea.pk.partner.dal.Job
 import com.bykea.pk.partner.dal.source.JobsDataSource
 import com.bykea.pk.partner.dal.source.JobsRepository
 import com.bykea.pk.partner.dal.source.pref.AppPref
+import com.bykea.pk.partner.repositories.UserDataHandler
+import com.bykea.pk.partner.repositories.UserRepository
 import com.bykea.pk.partner.ui.common.Event
 import com.bykea.pk.partner.ui.helpers.AppPreferences
 import com.bykea.pk.partner.utils.Constants
+import com.bykea.pk.partner.utils.Constants.ApiError.BOOKING_ALREADY_TAKEN
+import com.bykea.pk.partner.utils.Constants.ApiError.DRIVER_ACCOUNT_BLOCKED_BY_ADMIN
+import com.bykea.pk.partner.utils.Constants.DIGIT_ZERO
 import com.bykea.pk.partner.utils.Dialogs
 import com.bykea.pk.partner.utils.Util
 import com.bykea.pk.partner.utils.Utils
 import com.google.android.gms.maps.model.LatLng
+import org.apache.commons.lang3.StringUtils
 
 /**
  * The ViewModel used in [JobDetailActivity].
@@ -25,6 +31,14 @@ import com.google.android.gms.maps.model.LatLng
  * @Author: Yousuf Sohail
  */
 class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel(), JobsDataSource.GetJobRequestCallback, JobsDataSource.AcceptJobRequestCallback {
+
+    /**
+     * this property will contain the formatted name
+     * {including sender name + customer name (who created the booking)}
+     */
+    private val _formattedSenderName = MutableLiveData<String>().apply { value = StringUtils.EMPTY }
+    val formattedSenderName: LiveData<String>
+        get() = _formattedSenderName
 
     private val _currentLatLng = MutableLiveData<LatLng>()
     val currentLatLng: LiveData<LatLng>
@@ -65,6 +79,10 @@ class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel
     val showCOD: LiveData<Boolean>
         get() = _showCOD
 
+    private val _driverBlockedByAdmin = MutableLiveData<Boolean>().apply { value = false }
+    val driverBlockedByAdmin: MutableLiveData<Boolean>
+        get() = _driverBlockedByAdmin
+
     /**
      * Start the ViewModel by fetching the [Job] id
      *
@@ -82,7 +100,7 @@ class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel
      */
     fun accept() {
         job.value?.let {
-            jobsRepository.pickJob(it, this)
+            jobsRepository.pickJob(it, false, this)
         }
     }
 
@@ -101,6 +119,17 @@ class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel
     private fun renderDetails(job: Job?) {
         this._job.value = job
         _isDataAvailable.value = job != null
+        var formattedName = StringUtils.EMPTY
+        job?.sender?.name?.let {
+            formattedName = it
+            job.customer_name?.let {
+                //only show both when they are not same
+                if (!job.customer_name.equals(job.sender?.name, ignoreCase = true)) {
+                    formattedName = String.format(DriverApp.getContext().getString(R.string.formatted_name), job.sender?.name, job.customer_name)
+                }
+            }
+        }
+        this._formattedSenderName.value = formattedName
     }
 
     override fun onJobLoaded(job: Job) {
@@ -132,12 +161,26 @@ class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel
         _acceptBookingCommand.value = Event(Unit)
     }
 
-    override fun onJobRequestAcceptFailed(code: Int, message: String?) {
+    override fun onJobRequestAcceptFailed(code: Int, subCode: Int?, message: String?) {
         Dialogs.INSTANCE.dismissDialog()
         if (code == 422) {
-            _bookingTakenCommand.value = Event(Unit)
+            if (subCode != null && subCode != DIGIT_ZERO) {
+                when (subCode) {
+                    BOOKING_ALREADY_TAKEN -> {
+                        _bookingTakenCommand.value = Event(Unit)
+                    }
+                    DRIVER_ACCOUNT_BLOCKED_BY_ADMIN -> {
+                        _driverBlockedByAdmin.value = true
+                    }
+                    else -> {
+                        Dialogs.INSTANCE.showToast(message)
+                    }
+                }
+            } else {
+                Dialogs.INSTANCE.showToast(message)
+            }
         } else {
-            showSnackbarMessage(R.string.error_try_again)
+            Dialogs.INSTANCE.showToast(DriverApp.getContext().getString(R.string.error_try_again))
         }
     }
 
@@ -148,6 +191,10 @@ class JobDetailViewModel(private val jobsRepository: JobsRepository) : ViewModel
      */
     fun showSnackbarMessage(@StringRes message: Int) {
         _snackbarText.value = Event(message)
+    }
+
+    fun getTripDetails(userDataHandler: UserDataHandler) {
+        UserRepository().requestRunningTrip(DriverApp.getContext(), userDataHandler)
     }
 
 }
