@@ -3,6 +3,8 @@ package com.bykea.pk.partner.utils.audio
 import android.content.Context
 import android.media.MediaPlayer
 import android.net.Uri
+import java.io.File
+import java.io.FileInputStream
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.TimeUnit
@@ -10,7 +12,8 @@ import java.util.concurrent.TimeUnit
 class MediaPlayerHolder(context: Context) {
 
     private val mContext: Context
-    private var mMediaPlayer: MediaPlayer? = null
+    internal var mMediaPlayer: MediaPlayer? = null
+        get() = field
     private var mResourceId: Int = 0
     private var mAudioUri: String? = null
     private var mPlaybackInfoListener: PlaybackInfoListener? = null
@@ -79,6 +82,16 @@ class MediaPlayerHolder(context: Context) {
         }
 
         initializeProgressCallback()
+    }
+
+    fun loadFile(file: File) {
+        initializeMediaPlayer()
+        mMediaPlayer?.setDataSource(FileInputStream(file).fd);
+        mMediaPlayer?.prepare()
+        initializeProgressCallback()
+    }
+
+    fun retreiveFileFromAmazon(url: String) {
 
     }
 
@@ -102,7 +115,6 @@ class MediaPlayerHolder(context: Context) {
     fun reset() {
         if (mMediaPlayer != null) {
             mMediaPlayer!!.reset()
-            loadUri(mAudioUri)
             if (mPlaybackInfoListener != null) {
                 mPlaybackInfoListener!!.onStateChanged(PlaybackInfoListener.State.RESET)
             }
@@ -133,10 +145,10 @@ class MediaPlayerHolder(context: Context) {
             mSeekbarPositionUpdateTask = Runnable { updateProgressCallbackTask() }
         }
         mExecutor!!.scheduleAtFixedRate(
-            mSeekbarPositionUpdateTask,
-            0,
-            PLAYBACK_POSITION_REFRESH_INTERVAL_MS.toLong(),
-            TimeUnit.MILLISECONDS
+                mSeekbarPositionUpdateTask,
+                0,
+                PLAYBACK_POSITION_REFRESH_INTERVAL_MS.toLong(),
+                TimeUnit.MILLISECONDS
         )
     }
 
