@@ -6,19 +6,19 @@ import com.bykea.pk.partner.dal.util.EMPTY_STRING
 object AppPref {
 
     fun getLat(pref: SharedPreferences): Double {
-        return pref.getString(LATITUDE, "0.0").toDouble()
+        return pref.getString(LATITUDE, "0.0")?.toDouble()!!
     }
 
     fun getLng(pref: SharedPreferences): Double {
-        return pref.getString(LONGITUDE, "0.0").toDouble()
+        return pref.getString(LONGITUDE, "0.0")?.toDouble()!!
     }
 
     fun getDriverId(pref: SharedPreferences): String {
-        return pref.getString(DRIVER_ID, EMPTY_STRING)
+        return pref.getString(DRIVER_ID, EMPTY_STRING)!!
     }
 
     fun getAccessToken(pref: SharedPreferences): String {
-        return pref.getString(ACCESS_TOKEN, EMPTY_STRING)
+        return pref.getString(ACCESS_TOKEN, EMPTY_STRING)!!
     }
 
     fun getIsCash(pref: SharedPreferences): Boolean {
