@@ -1,9 +1,13 @@
 package com.bykea.pk.partner.dal.source.remote
 
 import com.bykea.pk.partner.dal.source.remote.response.BaseResponse
+import com.bykea.pk.partner.dal.source.remote.response.BaseResponseError
 import com.bykea.pk.partner.dal.util.ERROR_PLEASE_TRY_AGAIN
 import com.bykea.pk.partner.dal.util.INTERNAL_SERVER_ERROR
+import com.bykea.pk.partner.dal.util.UNAUTHORIZED
+import com.bykea.pk.partner.dal.util.UNAUTHORIZED_BROADCAST
 import com.google.gson.Gson
+import org.greenrobot.eventbus.EventBus
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -22,9 +26,18 @@ interface Callback<T : BaseResponse> : Callback<T> {
             if (response.errorBody() != null) {
                 try {
                     val res = Gson().fromJson(response.errorBody()?.string(), BaseResponse::class.java)
-
-                    onFail(res.code, res.subcode, res.message)
-                    onFail(res.code, res.message)
+                    if (isUserAuthorized(res.code)) {
+                        res.error?.let {
+                            if (it is String) {
+                                onFail(res.code, res.subcode, res.message)
+                            } else {
+                                onFail(res.code, Gson().fromJson(Gson().toJson(it), BaseResponseError::class.java), res.message)
+                            }
+                        } ?: run {
+                            onFail(res.code, res.subcode, res.message)
+                        }
+                        onFail(res.code, res.message)
+                    }
                 } catch (e: Exception) {
                     e.printStackTrace()
                     onFail(INTERNAL_SERVER_ERROR, INTERNAL_SERVER_ERROR, ERROR_PLEASE_TRY_AGAIN)
@@ -75,4 +88,24 @@ interface Callback<T : BaseResponse> : Callback<T> {
      * @param message Server message
      */
     fun onFail(code: Int, subCode: Int?, message: String?) {}
+
+    /**
+     * Negative response callback; will be called on fail with server code and message
+     * @param code Server code
+     * @param errorBody Error Body
+     * @param message Server message
+     */
+    fun onFail(code: Int, errorBody: BaseResponseError?, message: String?) {}
+
+    /**
+     * Check If User Is Authorized Or Not
+     * @param code : Code received in response from server
+     */
+    fun isUserAuthorized(code: Int): Boolean {
+        if (code == UNAUTHORIZED) {
+            EventBus.getDefault().post(UNAUTHORIZED_BROADCAST)
+            return false
+        }
+        return true
+    }
 }
