@@ -6,14 +6,8 @@ import com.google.gson.annotations.SerializedName
  * Fare Estimate API response.
  */
 class FareEstimationResponse : BaseResponse() {
-
-    //#region Getter Setter
-    @SerializedName("mstr")
-    var estimatedFare: String? = null
-
-    @SerializedName("calldata")
-    var callData: CallData? = null
-    //endregion
+    @SerializedName("data")
+    var fareEstimateData: FareEstimateData? = null
 }
 
 
