@@ -13,11 +13,15 @@ data class Trip(
         val end_lat: String,
         val end_lng: String,
         val ended_at: String,
+        val rule_ids: ArrayList<String>,
         val eta: String,
         val invoice_id: String,
         val status: String,
         val trip_time: String,
-        val wait_mins: Int
+        val wait_mins: Int,
+        val start_lat: Double,
+        val start_lng: Double,
+        val start_address: String
 )
 
 data class Invoice(
