@@ -2,4 +2,4 @@ package com.bykea.pk.partner.dal.source.remote.response
 
 data class StartJobResponse(val data: StartJobResponseData) : BaseResponse()
 
-data class StartJobResponseData(val trip_id: String)
+data class StartJobResponseData(var trip_id: String, var batch_id: String)
